change in valid integer

+++ semaphore.hpp
@@ -6,7 +6,7 @@ class Semaphore
 private:
     std::mutex mutex_;
     std::condition_variable condition_;
-    unsigned long count_;
+    unsigned long long count_;
 
 public:
     Semaphore(){
